Change the way single errors are treated
There is a potential pitfall when using this library:

```
var errors error
...
errors = multierror.Append(errors, doSomething())
...
return errors
```

Users might expect that if `doSomething` returns nil, then this
function will also return nil, and it thus might be the cause of subtle
bugs.

This commit changes the behaviour of Append slightly:

There will be never an instance of multierror that has exactly one
error. Or in other words: there will never be a multi-error that is
single and not multi.

Which kinda makes sense if you think about it. Having a message like:

```
there are 1 error(s):
something wrong
```

Is mostly just noise.

Implementation-wise it's straight forward.